requirements.txt
gdown
click==8.1.7
fasttext-wheel==0.9.2
joblib==1.3.2
lightgbm==4.3.0
nltk==3.8.1
numpy==1.20.1
pandas==1.2.2
pybind11==2.11.1
python-dateutil==2.8.2
pytz==2024.1
razdel==0.5.0
regex==2023.12.25
scikit-learn==0.24.1
scipy==1.12.0
six==1.16.0
threadpoolctl==3.2.0
tqdm==4.66.2
tzdata==2024.1
